Move Cost Estimation to Separate Class
